+++ src/main/groovy/com/ullink/gradle/opencover/OpenCoverNUnitPlugin.groovy
@@ -8,39 +8,38 @@ import org.gradle.api.Task
 class OpenCoverNUnitPlugin implements Plugin<Project> {
     void apply(Project project) {
         project.tasks.withType(OpenCover).whenTaskAdded { OpenCover task ->
-            applyOpencoverNunitConventions(task, project)
+            applyNunitDefaults(task, project)
         }
         project.apply plugin: 'com.ullink.opencover'
     }
 
-    def applyOpencoverNunitConventions(OpenCover task, Project project) {
+    def static applyNunitDefaults(OpenCover task, Project project) {
         Task nunit = project.tasks.nunit
         task.doFirst {
             nunit.prepareExecute()
         }
+        task.targetExec.set(project.provider { nunit.nunitExec })
+        task.targetExecArgs.set(project.provider { nunit.commandArgs.collect { it.toString() } })
+        task.parallelForks.set(project.provider { nunit.parallelForks })
 
-        task.conventionMapping.map 'targetExec', { nunit.nunitExec }
-        task.conventionMapping.map 'targetExecArgs', { nunit.commandArgs }
-        task.conventionMapping.map 'parallelForks',{ nunit.parallelForks }
+        def intermediateNunitResultsPath = new File(nunit.getReportFolderImpl().toString(), 'intemediate-results-nunit')
 
-        def intermediateNunitResultsPath = new File(nunit.getReportFolderImpl().toString() , "intemediate-results-nunit")
-
-        task.conventionMapping.map 'parallelTargetExecArgs', {
+        task.parallelTargetExecArgs.set(project.provider {
             intermediateNunitResultsPath.mkdirs()
-            nunit.getTestInputAsList(nunit.where?.value).collect({nunit.buildCommandArgs(it,  new File (intermediateNunitResultsPath,  UUID.randomUUID().toString() + ".xml"))})
-
-        }
+            nunit.getTestInputAsList(nunit.where?.value).collect({nunit.buildCommandArgs(it, new File (intermediateNunitResultsPath, UUID.randomUUID().toString() + ".xml"))})
+        })
 
         task.doLast {
-            def listOfIntermediateFiles = intermediateNunitResultsPath.listFiles().toList()
-            if (!listOfIntermediateFiles?.empty){
-                new NUnitTestResultsMerger().merge(listOfIntermediateFiles, nunit.testReportPath)
-                intermediateNunitResultsPath.deleteDir()
+            if (task.parallelForks.get() && !task.parallelTargetExecArgs.empty()) {
+                def listOfIntermediateFiles = intermediateNunitResultsPath.listFiles()?.toList()
+                if (listOfIntermediateFiles){
+                    new NUnitTestResultsMerger().merge(listOfIntermediateFiles, nunit.testReportPath)
+                    intermediateNunitResultsPath.deleteDir()
+                }
             }
         }
 
         project.afterEvaluate {
-            task.dependsOn nunit.dependsOn
             task.inputsOutputsFrom(nunit)
         }
     }